+++ zcash_primitives/src/transaction/components/sapling/builder.rs
@@ -60,6 +60,13 @@ impl fmt::Display for Error {
     }
 }
 
+/// A trait that provides a minimized view of a Sapling input suitable for use in
+/// fee calculation.
+pub trait SaplingInput {
+    /// The value of the input being spent.
+    fn value(&self) -> Amount;
+}
+
 #[derive(Debug, Clone)]
 pub struct SpendDescriptionInfo {
     extsk: ExtendedSpendingKey,
@@ -69,8 +76,18 @@ pub struct SpendDescriptionInfo {
     merkle_path: MerklePath<Node>,
 }
 
+impl SaplingInput for SpendDescriptionInfo {
+    fn value(&self) -> Amount {
+        // An existing note to be spent must have a valid
+        // amount value.
+        Amount::from_u64(self.note.value).unwrap()
+    }
+}
+
+/// A struct containing the information required in order to construct a
+/// Sapling output to a transaction.
 #[derive(Clone)]
-struct SaplingOutput {
+pub struct SaplingOutput {
     /// `None` represents the `ovk = ⊥` case.
     ovk: Option<OutgoingViewingKey>,
     to: PaymentAddress,
@@ -108,6 +125,12 @@ impl SaplingOutput {
             note,
             memo,
         })
+    }
+
+    pub fn value(&self) -> Amount {
+        // this unwrap is safe because the note's value was initially
+        // constructed from an `Amount`.
+        Amount::from_u64(self.note.value).unwrap()
     }
 
     fn build<P: consensus::Parameters, Pr: TxProver, R: RngCore>(
@@ -213,7 +236,7 @@ impl Authorization for Unauthorized {
     type AuthSig = SpendDescriptionInfo;
 }
 
-impl<P: consensus::Parameters> SaplingBuilder<P> {
+impl<P> SaplingBuilder<P> {
     pub fn new(params: P, target_height: BlockHeight) -> Self {
         SaplingBuilder {
             params,
@@ -225,11 +248,24 @@ impl<P: consensus::Parameters> SaplingBuilder<P> {
         }
     }
 
+    /// Returns the list of Sapling inputs that will be consumed by the transaction being
+    /// constructed.
+    pub fn inputs(&self) -> &[impl SaplingInput] {
+        &self.spends
+    }
+
+    /// Returns the Sapling outputs that will be produced by the transaction being constructed
+    pub fn outputs(&self) -> &[SaplingOutput] {
+        &self.outputs
+    }
+
     /// Returns the net value represented by the spends and outputs added to this builder.
     pub fn value_balance(&self) -> Amount {
         self.value_balance
     }
+}
 
+impl<P: consensus::Parameters> SaplingBuilder<P> {
     /// Adds a Sapling note to be spent in this transaction.
     ///
     /// Returns an error if the given Merkle path does not have the same anchor as the
@@ -293,15 +329,6 @@ impl<P: consensus::Parameters> SaplingBuilder<P> {
         self.outputs.push(output);
 
         Ok(())
-    }
-
-    /// Send change to the specified change address. If no change address
-    /// was set, send change to the first Sapling address given as input.
-    pub fn get_candidate_change_address(&self) -> Option<(OutgoingViewingKey, PaymentAddress)> {
-        self.spends.first().and_then(|spend| {
-            PaymentAddress::from_parts(spend.diversifier, spend.note.pk_d)
-                .map(|addr| (spend.extsk.expsk.ovk, addr))
-        })
     }
 
     pub fn build<Pr: TxProver, R: RngCore>(