+++ zcash_primitives/src/transaction/mod.rs
@@ -1,6 +1,7 @@
 //! Structs and methods for handling Zcash transactions.
 pub mod builder;
 pub mod components;
+pub mod fees;
 pub mod sighash;
 pub mod sighash_v4;
 pub mod sighash_v5;